refactor(event-list): remove mocks for main
