+++ app/src/androidTest/java/com/android/unio/ui/event/EventCardTest.kt
@@ -6,11 +6,19 @@ import com.android.unio.model.association.Association
 import com.android.unio.model.event.Event
 import com.android.unio.model.event.EventType
 import com.android.unio.model.event.MockEventRepository
-import com.android.unio.model.event.PreviewEventViewModel
+import com.android.unio.model.firestore.emptyFirestoreReferenceList
 import com.android.unio.model.firestore.firestoreReferenceListWith
 import com.android.unio.model.map.Location
-import com.android.unio.model.user.MockUserRepository
+import com.android.unio.model.user.Interest
+import com.android.unio.model.user.Social
+import com.android.unio.model.user.User
+import com.android.unio.model.user.UserRepository
+import com.android.unio.model.user.UserRepositoryFirestore
+import com.android.unio.model.user.UserSocial
+import com.android.unio.model.user.UserViewModel
+import com.google.firebase.Firebase
 import com.google.firebase.Timestamp
+import com.google.firebase.firestore.firestore
 import java.util.*
 import org.junit.Rule
 import org.junit.Test
@@ -33,11 +41,11 @@ class EventCardTest {
           location = Location(0.0, 0.0, "Sample Location"),
           types = listOf(EventType.TRIP))
 
-  private val viewModel = PreviewEventViewModel(MockEventRepository(), MockUserRepository())
+  private val userViewModel = UserViewModel(UserRepositoryFirestore(Firebase.firestore), false)
 
   @Test
   fun testEventCardElementsExist() {
-    composeTestRule.setContent { EventCard(event = sampleEvent, viewModel = viewModel) }
+    composeTestRule.setContent { EventCard(event = sampleEvent, userViewModel = userViewModel) }
 
     composeTestRule
         .onNodeWithTag("event_EventTitle", useUnmergedTree = true)
@@ -72,11 +80,74 @@ class EventCardTest {
 
   @Test
   fun testImageFallbackDisplayed() {
-    composeTestRule.setContent { EventCard(event = sampleEvent, viewModel = viewModel) }
+    composeTestRule.setContent { EventCard(event = sampleEvent, userViewModel = userViewModel) }
 
     // Check if the fallback image is displayed when no image is provided
     composeTestRule
         .onNodeWithTag("event_EventImage", useUnmergedTree = true)
         .assertExists() // Fallback image exists when no image is provided
   }
+}
+
+class MockUserRepository : UserRepository {
+    private val mockUsers =
+        listOf(
+            User(
+                uid = "1",
+                email = "john.doe@example.com",
+                firstName = "John",
+                lastName = "Doe",
+                biography = "Just a regular guy.",
+                followedAssociations = Association.emptyFirestoreReferenceList(),
+                joinedAssociations = Association.emptyFirestoreReferenceList(),
+                savedEvents = Event.emptyFirestoreReferenceList(),
+                interests = listOf(Interest.SPORTS, Interest.MUSIC),
+                socials =
+                listOf(
+                    UserSocial(Social.INSTAGRAM, "john_doe"),
+                    UserSocial(Social.WHATSAPP, "john.doe.whatsapp")),
+                profilePicture = "https://example.com/profile_picture1.jpg",
+                hasProvidedAccountDetails = false),
+            User(
+                uid = "2",
+                email = "jane.smith@example.com",
+                firstName = "Jane",
+                lastName = "Smith",
+                biography = "Lover of arts and technology.",
+                followedAssociations = Association.emptyFirestoreReferenceList(),
+                joinedAssociations = Association.emptyFirestoreReferenceList(),
+                savedEvents = Event.emptyFirestoreReferenceList(),
+                interests = listOf(Interest.ART, Interest.TECHNOLOGY),
+                socials =
+                listOf(
+                    UserSocial(Social.TELEGRAM, "jane_smith"),
+                    UserSocial(Social.TELEGRAM, "jane.smith.telegram")),
+                profilePicture = "https://example.com/profile_picture2.jpg",
+                hasProvidedAccountDetails = false))
+
+    override fun init(onSuccess: () -> Unit) {
+        onSuccess()
+    }
+
+    override fun getUsers(onSuccess: (List<User>) -> Unit, onFailure: (Exception) -> Unit) {
+        onSuccess(mockUsers)
+    }
+
+    override fun getUserWithId(
+        id: String,
+        onSuccess: (User) -> Unit,
+        onFailure: (Exception) -> Unit
+    ) {
+        val user = mockUsers.find { it.uid == id }
+        if (user != null) {
+            onSuccess(user)
+        } else {
+            onFailure(Exception("User not found"))
+        }
+    }
+
+
+    override fun updateUser(user: User, onSuccess: () -> Unit, onFailure: (Exception) -> Unit) {
+        TODO("Not yet implemented")
+    }
 }
